+++ src/Symfony/Component/Intl/ResourceBundle/Writer/TextBundleWriter.php
@@ -16,8 +16,9 @@ use Symfony\Component\Intl\Exception\UnexpectedTypeException;
 /**
  * Writes .txt resource bundles.
  *
- * The resulting files can be converted to binary .res files using the
- * {@link \Symfony\Component\Intl\ResourceBundle\Transformer\BundleCompiler}.
+ * The resulting files can be converted to binary .res files using a
+ * {@link \Symfony\Component\Intl\ResourceBundle\Transformer\BundleCompilerInterface}
+ * implementation.
  *
  * @author Bernhard Schussek <bschussek@gmail.com>
  *